Register hook creation
